requirements.txt
nltk
tqdm
elasticsearch
python-dotenv
sentence-transformers
